разеделил на модули и добавил докстринги
